Feature: preserve currently chosen spell
This change has two implications. First, on multiplayer game load the
drawn spell will be preserved. Sadly, when choosing characters, the spell
from previous game might also be preserved. This is fine. The game
is quite buggy - spell hotkeys are preserved across games as well,
so we are not making it much worse.